use (portable) u64 for player rank

+++ ferrogallic_shared/src/api/game.rs
@@ -98,7 +98,7 @@ pub struct Player {
 impl Player {
     pub fn rankings<'a>(
         players: impl IntoIterator<Item = (&'a UserId, &'a Player)>,
-    ) -> impl Iterator<Item = (usize, UserId, &'a Player)> {
+    ) -> impl Iterator<Item = (u64, UserId, &'a Player)> {
         let mut players_by_score = players
             .into_iter()
             .map(|(uid, player)| (*uid, player))
@@ -110,7 +110,7 @@ impl Player {
                 if player.score == *prev_score {
                     Some((*prev_rank, uid, player))
                 } else {
-                    let rank = index + 1;
+                    let rank = index as u64 + 1;
                     *prev_score = player.score;
                     *prev_rank = rank;
                     Some((rank, uid, player))
